// src/common/components/ModalCounterparty/enums/modalType.enum.ts
export enum EModalType {
  Counterparty = 'counterparty',
  EditContractorLegal = 'editContractorLegal',
  EditContractorIndividual = 'editContractorIndividual',
  AddCounterparty = 'addCounterparty',
  ConfirmDeleteCounterparty = 'confirmDeleteCounterparty',
}

export enum EINNLengths {
  Legal = 10,
  Individual = 12,
}